add support for custom modifier separators
